removes human test mistakenly added to automated_tests
